feet : 이미지 파일 유효성 검사 기능, FileService 인스턴스 및 LoacalFileService 생성
