+++ src/main/java/com/example/bestme/config/ChatGPTConfig.java
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
 
+
 @Configuration
 public class ChatGPTConfig {
