+++ src/main/java/com/example/bestme/handler/OAuth2SuccessHandler.java
@@ -30,6 +30,6 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 //        String token = jwtProvider.create(userId);
 //        String token = jwtProvider.generateToken();
 
-        response.sendRedirect("http://localhost:3000/auth/oauth-response/" + "token" + "/3600");
+        response.sendRedirect("http://localhost:3000/auth/oauth-response/" + token + "/3600");
     }
 }
