+++ src/main/resources/application.properties
@@ -50,3 +50,7 @@ spring.security.oauth2.client.provider.kakao.token-uri=https://kauth.kakao.com/o
 spring.security.oauth2.client.provider.kakao.user-info-uri=https://kapi.kakao.com/v2/user/me
 spring.security.oauth2.client.provider.kakao.user-name-attribute=id
 
+# ChatGPT API
+chatgpt.api.url=https://api.openai.com/v1/chat/completions
+chatgpt.model=gpt-3.5-turbo
+#chatgpt.api.key= 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
